+++ src/Lykke.RabbitMqBroker/SubscriberDependencyInjectionExtensions.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using Autofac;
 using Lykke.RabbitMqBroker.Subscriber;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection.Extensions;
 using RabbitMQ.Client;
 
 namespace Lykke.RabbitMqBroker
@@ -255,27 +256,31 @@ namespace Lykke.RabbitMqBroker
         /// <summary>
         /// Adds a connection provider to the service collection.
         /// Available as <see cref="IConnectionProvider"/>.
+        /// Idempotent.
         /// </summary>
         /// <param name="services"></param>
         public static void AddRabbitMqConnectionProvider(this IServiceCollection services)
         {
-            services.AddSingleton<IAutorecoveringConnectionFactory, AutorecoveringConnectionFactory>();
-            services.AddSingleton<IConnectionProvider, ConnectionProvider>();
+            services.TryAddSingleton<IAutorecoveringConnectionFactory, AutorecoveringConnectionFactory>();
+            services.TryAddSingleton<IConnectionProvider, ConnectionProvider>();
         }
         
         /// <summary>
         /// Adds a connection provider to the container
         /// Available as <see cref="IConnectionProvider"/>.
+        /// Idempotent. 
         /// </summary>
         /// <param name="builder"></param>
         public static void AddRabbitMqConnectionProvider(this ContainerBuilder builder)
         {
             builder.RegisterType<AutorecoveringConnectionFactory>()
                 .As<IAutorecoveringConnectionFactory>()
-                .SingleInstance();
+                .SingleInstance()
+                .IfNotRegistered(typeof(IAutorecoveringConnectionFactory));
             builder.RegisterType<ConnectionProvider>()
                 .As<IConnectionProvider>()
-                .SingleInstance();
+                .SingleInstance()
+                .IfNotRegistered(typeof(IConnectionProvider));
         }
     }
 }